Crate askama_parser

askama_parser: template parser for the Askama templating engine

This crate contains the procedural macros used by the Askama templating engine.

Re-exports

pub use crate::expr::AssociatedItem;

pub use crate::expr::Expr;

pub use crate::expr::Filter;

pub use crate::expr::PathComponent;

pub use crate::expr::TyGenerics;

pub use crate::expr::TyGenericsKind;

pub use crate::node::LetValueOrBlock;

pub use crate::node::Node;

Modules

expr

node

Structs

Ast

CharLit

InnerSyntax

NamedTarget

ParseError

Parsed

Span

A location in &'a str

StrLit

Syntax

SyntaxBuilder

WithSpan

Struct used to wrap types with their associated “span” which is used when generating errors in the code generation.

Enums

CharPrefix

FloatKind

IntKind

Num

PathOrIdentifier

StrPrefix

Target

Functions

is_rust_keyword

strip_common

Returns the common parts of two paths.