Crate askama_parser

Re-exports

pub use crate::expr::Attr;

pub use crate::expr::Expr;

pub use crate::expr::Filter;

pub use crate::expr::TyGenerics;

pub use crate::node::Node;

Modules

ascii_str

expr

node

Structs

Ast

CharLit

InnerSyntax

ParseError

Parsed

Span

An location in &'a str

StrLit

Syntax

SyntaxBuilder

WithSpan

Struct used to wrap types with their associated “span” which is used when generating errors in the code generation.

Enums

CharPrefix

FloatKind

IntKind

Num

StrPrefix

Target

Constants

MAX_RUST_KEYWORD_LEN

MAX_RUST_RAW_KEYWORD_LEN

RUST_KEYWORDS

Functions

strip_common

Returns the common parts of two paths.